feat: Add socket event to get teachers' stages
This commit adds a new socket event in the server code to retrieve the stages associated with a specific teacher. The event is triggered when the client requests the stages for a particular teacher. The server queries the database for stages where the encadrant field matches the provided teacher ID. The retrieved stages are then sent back to the client as a callback response.

Code changes:
- Added a new socket event "getTeachersStage" in the server's index.js file.
- Implemented the event handler function to query the Stage model for stages associated with the provided teacher ID.
- Sent the retrieved stages back to the client as a callback response.

Recent user commits:
- chore: Remove unnecessary code for file decoding in etudiant.js
- feat: Add "classe" field to etudiantSchema in Server/models/etudiant.js
- chore: Update MongoDB connection string in Server/index.js

Recent repository commits:
- encadrant
- Merge branch 'main' of https://github.com/halimchoukani/PFE_Manager
- aa
- feat: Add socket event to get students and populate select options
- fix admin rout
- chore: Add classe selection to EtudiantRegister form
- Add route for /cin in Server/index.js and Server/routes/cin.js, and add CIN validation in Server/models/cin.js