feat: Update links for Encadrant and Admin login in Acceuil.jsx
This commit updates the links for Encadrant and Admin login in the Acceuil.jsx file. The previous links were pointing to the "/etudiant/login" route for both, which is incorrect. The links have been corrected to "/encadrant/login" for Encadrant and "/admin/login" for Admin.

Code changes:
- Updated the link for Encadrant login to "/encadrant/login" in Acceuil.jsx.
- Updated the link for Admin login to "/admin/login" in Acceuil.jsx.

Recent user commits:
- zddz
- poizhozebg
- fzhzkbjgzkbgz

Recent repository commits:
- feat: Retrieve stages with associated files for teachers
- Merge branch 'main' of https://github.com/halimchoukani/PFE_Manager
- aa
- feat: Add socket event to get students and populate select options
- fix admin rout
- chore: Add classe selection to EtudiantRegister form
- Add route for /cin in Server/index.js and Server/routes/cin.js, and add CIN validation in Server/models/cin.js